refactor: pkg dataapps: create api pkg
