--- internal/pkg/service/appsproxy/dataapps/api.go
@@ -1,1 +1,0 @@
-package dataapps
