+++ internal/pkg/service/appsproxy/dataapps/api/api.go
@@ -1,0 +1,4 @@
+// Package api provides an interface for data apps API.
+package api
+
+type API struct{}
