+++ internal/pkg/service/appsproxy/dataapps/api/error.go
@@ -1,4 +1,4 @@
-package dataapps
+package api
 
 import (
 	"fmt"
