+++ internal/pkg/service/appsproxy/dataapps/api/notify.go
@@ -1,4 +1,4 @@
-package dataapps
+package api
 
 import (
 	"time"
